One-Time Ingestion/requirements.txt
datetime
markdownify
azure-storage-blob
azure-identity
beautifulsoup4
urllib3
lxml
python-dotenv